flash: Extract flash-specific storage routines
